Extract shared ensure helpers into internal/common
Move EnsureSecret, EnsureNetworkAttachments, EnsureTopology, and
ConditionUpdater into internal/common. Switch nova and placement
controllers to use the shared helpers.
